All Base CLI methods now return the same object.
Out of all existing Base CLI methods, ```info``` and ```list``` were
the only ones that did not return a python dictionary containing
stdout, stderr and retcode. I have updated them so that they now
return the proper object. The ```existing``` is the sole exception to
the rule.

While refactoring some of the tests that were affected by the changes
above, I had to explicitly skip a test from test_fact as it requires
some more work to setup the system for Facts.