CLI create method should always return dictionary
For some resources, the create method for the CLI was returning a list
with the created object as a dictionary, but for some other resources
the result was a plain dictionary. This fix should guarantee that when
a resource is created via the CLI, stdout should always be a dictionary.