Fixes the xunit file generated for Subnet tests.
There is an issue with using the ddt module and passing data with a
dot (.) in it, and how nose will generate its test name. If you pass a
data value with a dot (e.g. '192.168.100.0'), the test name dynamically
generated by nose splits off the last dot and ends up being that last
value. In this case the test is named '0'. So we annotate the name of
the test so that instead of using the data as its name, it uses the
integer representation of the ID of a python object.

Instead of seeing something like this:

```xml
<testcase
classname="tests.cli.test_subnet.TestSubnet.test_update_success_ddt_{'dns-primary':
'192.168" name="100" time="4.841"/>
```

we now end up with:

```xml
<testcase classname="tests.cli.test_subnet.TestSubnet" name="test_update_success_ddt_test_update_subnet_140236878070048" time="4.841"/>
```