Updated Smart Proxy CLI test.
We only have one CLI test for smart proxy and it has been failing since
forever due to the fact that you cannot create a random proxy with a
random URL. To create real tests will require some pipping under the
hood. In the meantime, I have updated the single test to verify that
there is an error when trying to create a proxy with a random URL. I
have also commented out the import for Proxy to quiet down PEP8.