Added new 'stubbed' decorator for stubbed tests.
Right now, in order to skip tests that have been 'stubbed' and don't
have any code implementing it, we have to import unittest, the
NOT_IMPLEMENTED constant and deal with unittest x unittest2 imports,
which is messy and duplicated in many places.

Instead I propose the creation of this 'stubbed' decorator to simplify
the process described above. If merged I'll create a separate PR (or
amend this one is desired) so that all existing tests are refactored.