Removing broken CLI test for Host.
This test was badly merged (it uses API calls for what should be a CLI
test) and it is also obsolete as systems are now called Content Host.
I'd like to delete this test (and avoid a false failure from our builds)
until I add a new set of CLI tests for Content Hosts.