Fixing CLI for systems tests.
Systems are now called Content Hosts, so I have renamed and moved
modules accordingly for CLI as well as fixed all imports. Also,
content-host command does not require an organization-id for certain
actions, such as Info. List does not accept 'per-page' so you need to
pass {'per-page': False} when invoking this action.

All tests included here are blocked by the fact that content-host create
returns a numeric ID for a new host, but all operations for a
content-host require a UUID.