Expose the TimeOut argument for CLI commands.
There are some CLI actions, such as content view promotions, can take a
bit longer to return and the default timeout of 50 seconds, which was
not exposed, was not long enough for some of our tests.

I have exposed the (optional) timeout argument via CLI's base class.
Also, I have set a default timeout of 120 seconds for content view
publishing for this PR.