Use skip_if_bz_bug_open instead of bzbug
Decorators are executed when a function is defined. As a result, importing a
module containing containing functions decorated with `bzbug` is an extremely
slow process: each of those funciton definitions triggers communications with a
Bugzilla server. The `skip_if_bz_bug_open` function accomplishes the same goal
of skipping certain tests. However, it does not incur communication with a
Bugzilla server until the test function is actually run.