Make `make` more resilient to errors
Remove the `NOSETESTS=$(shell which nosetests)` line and all lines which
reference that variable. Instead, use the executable directly. This change makes
`make` more resilient to errors. If nose is not installed and the command `make
test-robottelo` is executed, and error such as the following is generated:

    c robottelo.properties tests/robottelo/
    make: c: Command not found
    make: [test-robottelo] Error 127 (ignored)

If the executable is used directly, a much more helpful error is generated:

    nosetests -c robottelo.properties tests/robottelo/
    make: nosetests: Command not found
    make: *** [test-robottelo] Error 127

An examination of the make files used by Vim, OpenSSH and the Django docs shows
that none of them use `$(shell which ...)`, either.