Create some simple tests for path api/v2/models
Implement a few tests which check whether it's possible to GET or POST to the
path. Check for things like whether an HTTP 200 is generated when a GET request
is issued, that sane sets of data are returned, and that an HTTP 401 is
generated when an unauthorized HTTP request is made.