Update Makefile
Add a new 'help' recipe to the Makefile. Make this recipe the default. This
recipe lists out all available recipes and gives a brief description of their
effects.

Add a new 'test-docstrings' recipe to the Makefile. This recipe calls
`testimony`, just as Travis does right now.

Update the `.PHONY` special target. Wrap the line so as to be readable on
narrower editors.

Add comments to the makefile to visually separate its sections.