Make Entity.Meta.api_path simpler
Each entity's Meta.api_path is currently a tuple in this form:

    api_path = (
        '/api/v2/this_entity/',
        '/api/v2/this_entity/alternative_path',
    )

However, entities offer absolutely no support for intelligently choosing among
these URLs, which leads to confusing and ugly client code like this:

    return self.Meta.api_path[0]

This is a bad state of affairs. Convert each `api_path` to a simpler form:

    api_path = '/some/authoritative/path'

And trim down client code to deal with these changes.