Add instance method Entity.path
By default, this method returns the path used to access all entities of its
kind, but one can also ask for an individual entity's path.

    Organization().path() => 'example.com/api/v2/organizations'
    Organization().path('this') => NoSuchPathError
    Organization(id=25).path() => 'example.com/api/v2/organizations/25'
    Organization(id=25).path('all') => 'example.com/api/v2/organizations'

If one asks for a path which cannot be constructed, an exception is raised.

Add tests for this new method, and fix the setUp/tearDown methods for some
existing tests.