Perform all actions as new user.
This PR updates the existing CLI smoke test and asserts that all actions
are performed as the new user created at the start of the test. Prior to
this change, only the ``create`` actions were being performed as the new
user, and all the other actions were being performed as the default
``admin`` user.