Make DDT and skip_if_bug_open play nicely together
When the `@ddt` decorator is applied to a `unittest.TestCase` subclass, the
decorator modifies the all of that test case's methods. This is by design, and it is
correct behaviour. Unfortunately, the `skip_if_bug_open` decorator is a class,
so applying the decorator to a test method makes it... not a method. This
inadverdently causes the test to be skipped completely.

Fix issue #1158. Prevent test methods from disappearing.