Update LifecycleEnvironment field.
`entities.LifecycleEnvironment` has a `prior` foreignkey that points to
another instance of a `entities.LifecycleEnvironment`, which in turn is
renamed to `prior_id` before any actions can be performed. However, the
`Katello` API expects that this attribute will remain unchanged, so this
PR makes sure that `prior` will remain this named to avoid issues.