Revert changes to smoke test suite.
Just recently I updated the smoke test suite to go from
'Default_Organization' and 'Default_Location' to 'Default Organization'
and 'Default Location' respectively (note the absence of the
underscore). Well, turns out that only upstream `Katello` code applies
to this change but not downstream code. So sadly, this PR reverts these
changes and we will handle the upstream tests differently.