Remove arg from Factory.{attributes,build,create}
Remove the `fields` argument from methods `Factory.attributes`, `Factory.build`
and `Factory.create`. These three methods largely duplicated the functionality
provided by entities. This is bad in an of itself, because it meant that the
API-related code was not orthogonal:

    SomeEntity(field=value).create()
    SomeEntity().create(fields={'field': value})

However, the arg is also worth dropping because it makes the user vulnerable to
trivial errors. For example, consider these two statements:

    Organization(not_a_field=value).create()
    Organization().create(fields={'not_a_field': value})

Only one of them will raise an error, though both are clearly wrong.