Removed Capsule creation from smoke tests.
Our automation showed that having a `fake` Capsule created (as it was
being done by the CLI smoke tests) was causing a side effect with the
creation of `Repositories` and raising the following error:

::

  ``Action not allowed for the default capsule.``

This pull request removes the capsule creation and usage from existing
CLI smoke tests as well as updates docstrings mentioning the need for
its creation and use everywhere.

This pull request has been tested and it passed with flying colors on `qetello04`.