Make Repository.read work correctly
Add `EntityReadMixin` to the `Repository` class. Override the `read` method,
allowing it to deal with the fact that issuing an HTTP GET to the API returns a
"product" hash containing an ID, but not a "product_id" integer as expected.

Test this change.

Update `EntityReadMixin` to deal with the fact that `OneToOneField`s and
`OneToManyField`s may contain either classes or strings naming classes.