Add API tests for host collections
Update the `HostCollection` class definition: add mixins for creation, reading
and deletion. Remove the `system_uuids` attribute: no information of the sort is
provided when reading a host collection, and when creating or updating a host
collection, the attribute seems to conflict with the `system_ids` attribute.

Thoroughly test the relevant API endpoints.