Add clients provisioning feature
The clients provisioning feature will allow create clients to be used on
tests that needs another machine to consume some content.

Also this improve the ssh module to allow specify the machine where to
run a command. Improved logging to show what machine the command was
executed (inspired by Fabric).

Finally was added tests for the new robotello.vm module and properly
added entries for added modules in the docs.

Bonus: refactored logging on the ssh module.