Trim down ForemanTask, test ForemanTask API
Remove the custom method `ForemanTask.read`. Make `ForemanTask` inherit from
`EntityReadMixin`. `ForemanTask.read_json` performs the same task that the old
`ForemanTask.read` method did, and no code was using the old method.

Add a test targeting bugzilla bug 1131702. The test asserts that when a
non-existent foreman task is read, an HTTP 4XX or 5XX error is received. The
test fails, so it is skipped appropriately.

Add this new test module to the API docs.