Manifests-UI: Enabling the subscriptions and sync related tests.
Manifests can now be uploaded via the UI using the `clone` function
from `robottelo.common.manifests` module.

The `clone` function now uses by default, the `key_file` path  and
the `manifest_template` path which is specified in the,
robottelo.properties file.

Also tests related to enabling of RedHat Repos and Syncing these
Repos have also been fixed..

We are using RedHat Manifests, cloning them and just changing the
UUID in consumer.json file, so as to enable it to import in
different Organizations.