Convert old API tests for users
Delete module `test_users`. Rewrite module `test_user_v2`, and drop in a test
that thoroughly tests whether it is possible to create and delete a user with a
variety of attributes.

Update the definition of the `User` class. Set some better limits on how long
the fields `firstname`, `lastname` and `login` can be.