Add log messages for virtual display feature
Sometimes the window manager process hangs and the test suite stuck. To
better find the cause was added logging messages to show when a virtual
display and a window manager is created, will be stopped and stopped.
The log messages will help debug the hanging problem and also will
provide more information about what is going on on.

Bonus: docstring formatting for robottelo.test module.