(API) Fix up the "OperatingSystem" entity.
Give the `OperatingSystem` entity three new attributes:

* architecture
* medium
* ptable

Override `OperatingSystem.create` so that these attributes can be correctly
specified. Write new API tests for the "architecture" and "ptable" attributes.
Do not create a test for the "medium" attribute, as it is not currently possible
to create a medium using the API tools.