Fix for the UI boundary tests.
Boundary tests need a fixed length to test whether the entity
creation beyond 255 chars fails. It is here where we do require
an option to specify a fixed length.

so now unless we specify a length for the string it will continue
to use a random length for non-boundary tests.

Added sphinx directives