Avoid 404 on faster machines and connections
Add a retry with 1 second sleep before it when get a 404 on reading an
entity. Some API tests are failing due to 404, but when accessing the
URL the entity is returned without any problems. This was not possible
to reproduce locally, had to run on the Jenkins slave which is faster
and have a faster connection to the Satellite server.