Fix robottelo.api.client
The Python requests library has changed upstream. Make module
`robottelo.api.client` match the upstream changes.

Update the tests for this module. Use the mock library. This allows mock classes
to be dropped, manual patching/unpatching to be dropped, and function calls to
be inspected at a more granual level.