Update smoke test for available URLs
Rewrite `TestAvailableURLs.test_get_links` to match the response currently
generated by the API.

[BZ 1105773](https://bugzilla.redhat.com/show_bug.cgi?id=1105773) was marked as
VERIFIED on the 10th. As a result, the `skip_if_bug_open` decorator now allows
the test to run. The test expects a flat list of URLs from the API:

    u'/katello/api/content_views/:content_view_id/content_view_puppet_modules',
    u'/katello/api/content_views/:content_view_id/filters',
    u'/katello/api/content_view_versions',
    u'/katello/api/environments',
    u'/katello/api/errata',
    u'/katello/api/gpg_keys',
    u'/katello/api/host_collections',
    u'/katello/api/organizations',
    u'/katello/api/organizations/:organization_id/content_views',
    …

The test was created quite some time back, and in the meantime the API has been
changed. The API actually returns data in this format:

    {u'content_views': {
        u'…': 'u/katello/api/content_views/:id',
        u'…': 'u/katello/api/content_views/:id',
        u'…': 'u/katello/api/content_views/:id',
        u'…': 'u/katello/api/content_views/:id/available_puppet_module_names',
        u'…': 'u/katello/api/content_views/:id/available_puppet_modules',
        u'…': 'u/katello/api/content_views/:id/copy',
        u'…': 'u/katello/api/content_views/:id/environments/:environment_id',
        u'…': 'u/katello/api/content_views/:id/history',
        u'…': 'u/katello/api/content_views/:id/publish',
        u'…': 'u/katello/api/content_views/:id/remove',
        u'…': 'u/katello/api/organizations/:organization_id/content_views',
        u'…': 'u/katello/api/organizations/:organization_id/content_views',
    }, …}

The keys in the outer dict are names like "content_views" or "architectures".
The keys in the inner dict are prose descriptions of corresponding paths.

Fix #1555.