Make the readme a real file, not a symlink
The readme should render correctly when processed by rst2html, displayed on
GitHub, or read as a standalone file. In contrast, the main body of
documentation must be compiled before being read. Split the readme and main body
of documentation apart.

Add a table of contents to `docs/index.rst`.

Fix a random error in `robottelo/orm.py`.