set_context function in ui/factory no longer depends on "@" char.
We now fetch both organization and location separately, hence no
longer we need to deal with "@" char. This is undesired as org and
location both could contains "@" char and complicate things.

Also no need to explicitly set the navigation again when creating
second organization, ui/factory should take care of it when the
force_context=True.