Drop method update_dictionary
Method `update_dictionary` behaves almost exactly like the `update` method from
the standard library. Consider these two lines of code:

    d1.update(d2)
    update_dictionary(d1, d2)

In the first case, `d1` will receive all keys from `d2`:

    >>> d1 = {1: 10, 2: 20}
    >>> d2 = {2: 25, 3: 30}
    >>> d1.update(d2)
    >>> d1
    {1: 10, 2: 25, 3: 30}

In the second case, `d1` will receive only intersecting keys from `d2`:

    >>> d1 = {1: 10, 2: 20}
    >>> d2 = {2: 25, 3: 30}
    >>> update_dictionary(d1, d2)
    >>> helpers.update_dictionary(d1, d2)
    {1: 10, 2: 25}

`update_dictionary` is used in two modules:

* robottelo.cli.factory
* robottelo.ui.factory

In the UI factory, the usage of `update_dictionary` is a no-op, as `update` is
called immediately afterwards. In the CLI factory, the usage of
`update_dictionary` appears to be useless, though this is not a logical
certainty.