Added Submition validation for sync_plan tests.
Upon submission of any page, there is a chance that the validation
of data may fail and hence the entity creation too fails.

The idea is to check for the submit/create/save locators after the
submission of page. If these locators still exist on the page then
the validation fails.

Added a submit_validate flag so that we have a choice to validate
whether submission was successful. This needs to be used for all
the negative tests, as submission would always fail for them.

submit_validate flag is set to True by default.
Needs to be set to false for all negative tests.